Improved spoken css recognition
Sometimes Annyang.js returns `colon:separated` words, as if it knew someone was speaking CSS.

I'd noticed this previously, and was watching out for cases that ended with `px` or `%`, but since there are (obviously) other cases, I've made that match case more general now.

This fixes #1, thanks for the report @kumavis!